Update NillaEventPrice.java
Nilaa’s Catering provides meals for parties and special events. Write a program that prompts the user for the number of guests attending an event and then computes the total price, which is $35 per person. 	Display the number of guests, price per guest, and total price. 
Also display a message that indicates true or false depending on whether the job is classified as a large event—an event with 50 or more guests. 
		Save the file as NilaaEventPrice.java.